MI-Fitness/app/requirements.txt
flask==1.1.2
flask-bcrypt==0.7.1
flask-login==0.5.0
flask-sqlalchemy==2.5.1
flask-wtf==0.14.3
wtforms==2.3.3
email-validator==1.1.3
sklearn
lightgbm
numpy